app/pyproject.toml
[tool.poetry]
name = "devops-adventure-cafezinho"
version = "0.1.0"
description = ""
authors = ["LstCenturion <allan.yoshioka@gmail.com>"]
readme = "README.md"
packages = [
    { include = ".src" },
    { include = ".db" }
]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.32.0"
fastapi = "^0.115.3"
requests = "^2.32.3"
python-multipart = "^0.0.12"
asyncpg = "^0.30.0"
databases = "^0.9.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
